Implement setting drive strength

+++ esphome/components/pca9554/pca9554.h
@@ -21,7 +21,7 @@ class PCA9554Component : public Component,
   virtual void pin_mode(uint8_t pin, gpio::Flags flags);  // TODO: do we need virtual here?
   void set_latch(uint8_t pin, bool latch_state);
   void set_interrupt(uint8_t pin, bool interrupt_state);
-  void set_drive_strength(uint8_t pin, uint8_t strength_to_set);
+  void set_drive_strength(uint8_t pin, uint16_t strength_to_set);
   void set_open_drain(uint8_t bank_to_set, bool state_open_drain);
   uint16_t get_interrupt_status();
 
@@ -33,7 +33,7 @@ class PCA9554Component : public Component,
   void set_interrupt_pin(InternalGPIOPin *pin) { this->interrupt_pin_ = pin; }
 
   // test function, delete this later
-  // void do_stuff();
+  //void do_stuff();
 
  protected:
   static void IRAM_ATTR gpio_intr(PCA9554Component *arg);
